Clean hero sections, rework home page and update header links order
